backend/energy-dashboard/requirements.txt
streamlit>=1.33.0
pandas>=2.2.0
openpyxl>=3.1.2
plotly>=5.20.0
jinja2>=3.1.3
python-dotenv>=1.0.1
pyyaml>=6.0.1
schedule>=1.2.1
